requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
bokeh==3.2.2
    # via -r requirements.in
contourpy==1.1.0
    # via bokeh
jinja2==3.1.2
    # via bokeh
markupsafe==2.1.3
    # via jinja2
numpy==1.25.2
    # via
    #   -r requirements.in
    #   bokeh
    #   contourpy
    #   pandas
packaging==23.1
    # via bokeh
pandas==2.1.0
    # via bokeh
pillow==9.5.0
    # via bokeh
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via bokeh
six==1.16.0
    # via python-dateutil
tornado==6.3.3
    # via bokeh
tzdata==2023.3
    # via pandas
xyzservices==2023.7.0
    # via bokeh
